Authenticate requests

+++ TodoApp/routers/todos.py
@@ -5,6 +5,7 @@ from sqlalchemy.orm import Session
 from typing import Annotated
 from pydantic import BaseModel, Field
 from starlette import status
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -16,6 +17,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -24,25 +26,39 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get('/todos', status_code = status.HTTP_200_OK)
-async def get_all_todos(db: db_dependency):
-    return db.query(Todo).all()
+async def get_all_todos(user: user_dependency, 
+                        db: db_dependency):
+    return db.query(Todo).filter(Todo.owner_id == user.get('id')).all()
 
 @router.get('/todos/{todo_id}', status_code = status.HTTP_200_OK)
-async def get_todo_with_id(db: db_dependency, todo_id: int = Path(gt = 0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+async def get_todo_with_id(user: user_dependency, 
+                           db: db_dependency, 
+                           todo_id: int = Path(gt = 0)):
+    if user is None:
+        raise HTTPException(status_code = status.HTTP_401_UNAUTHORIZED, details = 'User authentication failed')
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is not None:
         return todo
     raise HTTPException(status_code = 400, detail = "Todo not found")
 
 @router.post('/todos', status_code = status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo = Todo(**todo_request.model_dump())
+async def create_todo(user: user_dependency, 
+                      db: db_dependency, 
+                      todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code = status.HTTP_401_UNAUTHORIZED, details = 'User authentication failed')
+    todo = Todo(**todo_request.model_dump(), owner_id = user.get('id'))
     db.add(todo)
     db.commit()
 
 @router.put('/todos/{todo_id}', status_code = status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt = 0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+async def update_todo(user: user_dependency, 
+                      db: db_dependency, 
+                      todo_request: TodoRequest, 
+                      todo_id: int = Path(gt = 0)):
+    if user is None:
+        raise HTTPException(status_code = status.HTTP_401_UNAUTHORIZED, details = 'User authentication failed')
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is None:
         raise HTTPException(status_code = 400, detail = "Todo not found")
     
@@ -55,10 +71,14 @@ async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int
     db.commit()
 
 @router.delete('/todos/{todo_id}', status_code = status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+async def delete_todo(user: user_dependency, 
+                      db: db_dependency, 
+                      todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code = status.HTTP_401_UNAUTHORIZED, details = 'User authentication failed')
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is None:
         raise HTTPException(status_code = 400, detail = "Todo not found")
     
-    db.query(Todo).filter(Todo.id == todo_id).delete()
+    db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).delete()
     db.commit()
